Move createJsonReponse into separate file
- We will move this to typescript/react-native-verus module later
