+++ src/utils/api/channels/dlight/requests/getInfo.js
@@ -1,27 +1,7 @@
 import { makeDlightRequest } from '../callCreators'
 import { DLIGHT_INFO } from '../../../../constants/dlightConstants'
 import { getSynchronizerInstance } from 'react-native-verus'
-
-
-export const createJsonRpcResponse = (id, result, error) => {
-  const responseData = {
-    jsonrpc: "2.0",
-    id: id,
-  }
-
-  if (error) {
-    responseData.error = error;
-  } else {
-    responseData.result = result;
-  }
-  const response = new Response(JSON.stringify(responseData), {
-    status: 200,
-    headers: {
-      'Content-Type': 'application/json'
-    },
-  });
-  return response.json()
-}
+import { createJsonRpcResponse } from './jsonResponse'
 
 // Get the syncing status/information about a blockchain
 export const getInfo = (coinId, accountHash, coinProto) => {
